employee/serializers.py
from rest_framework import serializers
from .models import EmployeeMaster

class EmployeeMasterSerializer(serializers.ModelSerializer):
    class Meta:
        model = EmployeeMaster
        fields = '__all__'
