requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-import-export==3.0.1
django-jazzmin==2.6.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.13.2
pillow-10.2.0
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
djangorestframework
